+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Update context (#270)
 - Added theme object, notifySuccess & notifyError (#270)
+- Update dependencies and Actions (#271)
 
 ## 4.3.0 (2023-12-25)