+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Added theme object, notifySuccess & notifyError (#270)
 - Update dependencies and Actions (#271)
 
+### Bug fixes
+
+- Fix statusColor for Row (#280)
+
 ## 4.3.0 (2023-12-25)
 
 ### Features / Enhancements