+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 4.4.0 (IN PROGRESS)
+## 4.4.0 (2024-03-06)
 
 ### Features / Enhancements
 
@@ -10,6 +10,7 @@
 - Replace custom code parameters with Code Parameters Builder (#285)
 - Update CSS class for the Panel instead of a Row (#272)
 - Update Editor auto height from fixed value (#278)
+- Disable unitScale and fieldMinMax standard panel options (#286)
 
 ### Bug fixes